OutfitFlowApi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
rembg==2.0.56
pillow==10.1.0
